refactor(db): Phase 1 complete — SQL extracted from all modules
Baseline: 432 → 9 SQL calls outside db/ (97.9% extracted).
The 9 residual are cursor-sharing orchestration blocks in bliss.py
(read-only radio composition that shares a transaction across
multiple helper calls). These will resolve naturally when
transaction_scope() evolves in Phase 2.

API routers migrated (all at 0):
  analytics.py (45), browse_media.py (45), browse_artist.py (31),
  subsonic.py (25), me.py (16), browse_shared.py (5), setup.py (3),
  auth.py (3), artwork.py (3), acquisition.py (3), playlist_utils.py (3),
  _deps.py (3), system_playlists.py (1)

Worker handlers migrated (all at 0):
  enrichment.py (25), migration.py (20), analysis.py (20),
  management.py (19), artwork.py (5), acquisition.py (5),
  integrations.py (3), __init__.py (3)

Standalone modules migrated (all at 0 except bliss.py residual):
  bliss.py (36→8), library_sync.py (24→0), analysis_daemon.py (15→0),
  enrichment.py (3→0)

New db modules created:
  db/queries/analytics.py     — 30 functions (overview, stats, insights)
  db/queries/user.py          — 8 functions (feed, upcoming, scrobble)
  db/queries/subsonic.py      — 14 functions (full Subsonic protocol)
  db/queries/browse_media.py  — 16 functions (search, favorites, tracks)
  db/queries/browse_artist.py — 15 functions (filters, pagination, genres)
  db/queries/bliss.py         — 18 functions (vectors, similarity, radio)
  db/jobs/sync.py             — 8 functions (upserts, counts, merges)
  db/jobs/analysis.py         — 14 functions (claim, status, vectors)
  db/jobs/enrichment.py       — 13 functions (MBID, popularity, covers)
  db/jobs/migration.py        — 8 functions (path updates, listings)
  db/jobs/management.py       — 4 functions (rename, MBID apply)
  db/jobs/artwork.py          — 2 functions (photo/cover flags)
  db/jobs/acquisition.py      — 1 function (release dates)
  db/jobs/integrations.py     — 1 function (similar_json)
  db/jobs/tasks.py            — 1 function (task status)

Existing modules extended:
  db/library.py, db/auth.py, db/playlists.py, db/management.py,
  db/cache.py, db/genres.py, db/shows.py, db/tasks.py

Tests: 252 passing, 0 failures. 7 test files updated with new
mock paths (test_api, test_auth, test_setup, test_explore_contracts,
test_library_sync, test_radio_contracts).
