+++ app/crate/api/browse_media.py
@@ -7,7 +7,27 @@ from fastapi.responses import JSONResponse, Response
 from crate.api._deps import enrich_radio_tracks as _enrich_radio_tracks, library_path, safe_path
 from crate.api.auth import _require_auth
 from crate.api.browse_shared import _YEAR_PREFIX_RE, fs_search, has_library_data
-from crate.db import get_cache, get_db_ctx, set_cache
+from crate.db import get_cache, set_cache
+from crate.db.queries.browse_media import (
+    add_favorite,
+    count_mood_tracks,
+    find_track_id_by_path,
+    get_mood_tracks,
+    get_track_album_genres,
+    get_track_artist_genres,
+    get_track_exists,
+    get_track_id_by_storage_id,
+    get_track_info_cols,
+    get_track_info_cols_by_path,
+    get_track_info_cols_by_storage_id,
+    get_track_path,
+    get_track_path_by_storage_id,
+    list_favorites,
+    remove_favorite,
+    search_albums,
+    search_artists,
+    search_tracks,
+)
 
 log = logging.getLogger(__name__)
 
@@ -28,46 +48,9 @@ def api_search(request: Request, q: str = "", limit: int = 20):
         return result
 
     like = f"%{q_stripped}%"
-    with get_db_ctx() as cur:
-        cur.execute(
-            """
-            SELECT id, slug, name, album_count, has_photo
-            FROM library_artists
-            WHERE name ILIKE %s
-            ORDER BY listeners DESC NULLS LAST, album_count DESC, name ASC
-            LIMIT %s
-            """,
-            (like, capped_limit),
-        )
-        artist_rows = cur.fetchall()
-        cur.execute(
-            """
-            SELECT a.id, a.slug, a.artist, a.name, a.year, a.has_cover,
-                   ar.id AS artist_id, ar.slug AS artist_slug
-            FROM library_albums a
-            LEFT JOIN library_artists ar ON ar.name = a.artist
-            WHERE a.name ILIKE %s OR a.artist ILIKE %s
-            ORDER BY year DESC NULLS LAST, name ASC
-            LIMIT %s
-            """,
-            (like, like, capped_limit),
-        )
-        album_rows = cur.fetchall()
-        cur.execute(
-            """
-            SELECT t.id, t.storage_id, t.slug, t.title, t.artist, a.id AS album_id, a.slug AS album_slug,
-                   a.name AS album, ar.id AS artist_id, ar.slug AS artist_slug,
-                   t.path, t.duration
-            FROM library_tracks t
-            JOIN library_albums a ON t.album_id = a.id
-            LEFT JOIN library_artists ar ON ar.name = t.artist
-            WHERE t.title ILIKE %s OR t.artist ILIKE %s OR a.name ILIKE %s
-            ORDER BY t.title ASC
-            LIMIT %s
-            """,
-            (like, like, like, capped_limit),
-        )
-        track_rows = cur.fetchall()
+    artist_rows = search_artists(like, capped_limit)
+    album_rows = search_albums(like, capped_limit)
+    track_rows = search_tracks(like, capped_limit)
 
     artists = [
         {
@@ -115,10 +98,7 @@ def api_search(request: Request, q: str = "", limit: int = 20):
 @router.get("/api/favorites")
 def api_favorites_list(request: Request):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT item_type, item_id, created_at FROM favorites ORDER BY created_at DESC")
-        items = [dict(row) for row in cur.fetchall()]
-    return {"items": items}
+    return {"items": list_favorites()}
 
 
 @router.post("/api/favorites/add")
@@ -134,11 +114,7 @@ def api_favorites_add(request: Request, body: dict):
         return JSONResponse({"error": "type must be song, album, or artist"}, status_code=400)
 
     now = datetime.now(timezone.utc).isoformat()
-    with get_db_ctx() as cur:
-        cur.execute(
-            "INSERT INTO favorites (item_type, item_id, created_at) VALUES (%s, %s, %s) ON CONFLICT DO NOTHING",
-            (item_type, item_id, now),
-        )
+    add_favorite(item_type, item_id, now)
 
     return {"ok": True}
 
@@ -153,8 +129,7 @@ def api_favorites_remove(request: Request, body: dict):
     if item_type not in ("song", "album", "artist"):
         return JSONResponse({"error": "type must be song, album, or artist"}, status_code=400)
 
-    with get_db_ctx() as cur:
-        cur.execute("DELETE FROM favorites WHERE item_id = %s AND item_type = %s", (item_id, item_type))
+    remove_favorite(item_type, item_id)
 
     return {"ok": True}
 
@@ -172,10 +147,7 @@ def api_rate_track(request: Request, body: dict):
         return JSONResponse({"error": "Rating must be 0-5"}, status_code=400)
 
     if not track_id and track_path:
-        with get_db_ctx() as cur:
-            cur.execute("SELECT id FROM library_tracks WHERE path LIKE %s LIMIT 1", (f"%{track_path}",))
-            row = cur.fetchone()
-            track_id = row["id"] if row else None
+        track_id = find_track_id_by_path(track_path)
 
     if not track_id:
         return JSONResponse({"error": "Track not found"}, status_code=404)
@@ -232,9 +204,7 @@ def _serialize_track_info_row(row) -> dict:
 @router.get("/api/tracks/{track_id}/info")
 def api_track_info_by_id(request: Request, track_id: int):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute(f"SELECT {_TRACK_INFO_QUERY_COLS} FROM library_tracks WHERE id = %s", (track_id,))
-        row = cur.fetchone()
+    row = get_track_info_cols(track_id, _TRACK_INFO_QUERY_COLS)
     if not row:
         return Response(status_code=404)
     return _serialize_track_info_row(row)
@@ -243,9 +213,7 @@ def api_track_info_by_id(request: Request, track_id: int):
 @router.get("/api/tracks/by-storage/{storage_id}/info")
 def api_track_info_by_storage_id(request: Request, storage_id: str):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute(f"SELECT {_TRACK_INFO_QUERY_COLS} FROM library_tracks WHERE storage_id = %s", (storage_id,))
-        row = cur.fetchone()
+    row = get_track_info_cols_by_storage_id(storage_id, _TRACK_INFO_QUERY_COLS)
     if not row:
         return Response(status_code=404)
     return _serialize_track_info_row(row)
@@ -257,12 +225,7 @@ def api_track_info(request: Request, filepath: str):
     if filepath.startswith("/music/"):
         filepath = filepath[len("/music/") :]
 
-    with get_db_ctx() as cur:
-        cur.execute(
-            f"SELECT {_TRACK_INFO_QUERY_COLS} FROM library_tracks WHERE path LIKE %s LIMIT 1",
-            (f"%{filepath}",),
-        )
-        row = cur.fetchone()
+    row = get_track_info_cols_by_path(filepath, _TRACK_INFO_QUERY_COLS)
 
     if not row:
         return Response(status_code=404)
@@ -270,16 +233,6 @@ def api_track_info(request: Request, filepath: str):
 
 
 # ── EQ adaptive features ────────────────────────────────────────────
-#
-# Minimal subset of audio-analysis columns exposed for the client-side
-# adaptive equalizer heuristic. Keep this narrow: the front-end shouldn't
-# learn about every analysis column we persist, and this payload gets
-# requested on every track change when the user selects the "adaptive"
-# preset — a smaller JSON = fewer bytes on the wire and a clearer
-# contract.
-#
-# spectral_complexity already lives in [0, 1] (normalized centroid at
-# analysis time), so it doubles as a brightness indicator.
 
 _EQ_FEATURES_QUERY_COLS = (
     "energy, loudness, dynamic_range, spectral_complexity, "
@@ -292,9 +245,9 @@ def _serialize_eq_features(row) -> dict:
     data = dict(row)
     return {
         "energy": data.get("energy"),
-        "loudness": data.get("loudness"),              # LUFS, roughly -30..-6
-        "dynamicRange": data.get("dynamic_range"),     # dB crest-like
-        "brightness": data.get("spectral_complexity"), # normalized centroid 0..1
+        "loudness": data.get("loudness"),
+        "dynamicRange": data.get("dynamic_range"),
+        "brightness": data.get("spectral_complexity"),
         "danceability": data.get("danceability"),
         "valence": data.get("valence"),
         "acousticness": data.get("acousticness"),
@@ -305,12 +258,7 @@ def _serialize_eq_features(row) -> dict:
 @router.get("/api/tracks/{track_id}/eq-features")
 def api_eq_features_by_id(request: Request, track_id: int):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute(
-            f"SELECT {_EQ_FEATURES_QUERY_COLS} FROM library_tracks WHERE id = %s",
-            (track_id,),
-        )
-        row = cur.fetchone()
+    row = get_track_info_cols(track_id, _EQ_FEATURES_QUERY_COLS)
     if not row:
         return Response(status_code=404)
     return _serialize_eq_features(row)
@@ -319,26 +267,13 @@ def api_eq_features_by_id(request: Request, track_id: int):
 @router.get("/api/tracks/by-storage/{storage_id}/eq-features")
 def api_eq_features_by_storage_id(request: Request, storage_id: str):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute(
-            f"SELECT {_EQ_FEATURES_QUERY_COLS} FROM library_tracks WHERE storage_id = %s",
-            (storage_id,),
-        )
-        row = cur.fetchone()
+    row = get_track_info_cols_by_storage_id(storage_id, _EQ_FEATURES_QUERY_COLS)
     if not row:
         return Response(status_code=404)
     return _serialize_eq_features(row)
 
 
 # ── Track primary genre ─────────────────────────────────────────────
-#
-# Returns the dominant genre for a track, used by the "Genre Adaptive"
-# equalizer mode. We pick the highest-weight canonical genre from the
-# track's album; if no album genres exist (or none are canonical) we
-# fall back to the artist's genres. Non-canonical tags (raw Last.fm
-# noise like "blackened-post-sludge") are still returned if nothing
-# canonical is available, but `primary.canonical = false` tells the
-# client to skip preset lookup.
 
 def _pick_primary_genre(rows):
     """Prefer the highest-weight canonical genre; fall back to the
@@ -400,38 +335,14 @@ def _pick_primary_genre(rows):
     return canonical_pick or raw_pick
 
 
-def _resolve_track_genre(cur, track_id: int) -> dict | None:
-    cur.execute(
-        """
-        SELECT g.name, g.slug, ag.weight
-        FROM library_tracks t
-        JOIN album_genres ag ON ag.album_id = t.album_id
-        JOIN genres g ON g.id = ag.genre_id
-        WHERE t.id = %s
-        ORDER BY ag.weight DESC NULLS LAST, g.name ASC
-        LIMIT 10
-        """,
-        (track_id,),
-    )
-    album_rows = cur.fetchall()
+def _resolve_track_genre(track_id: int) -> dict | None:
+    album_rows = get_track_album_genres(track_id)
     picked = _pick_primary_genre(album_rows) if album_rows else None
     if picked:
         picked["source"] = "album"
         return picked
 
-    cur.execute(
-        """
-        SELECT g.name, g.slug, arg.weight
-        FROM library_tracks t
-        JOIN artist_genres arg ON arg.artist_name = t.artist
-        JOIN genres g ON g.id = arg.genre_id
-        WHERE t.id = %s
-        ORDER BY arg.weight DESC NULLS LAST, g.name ASC
-        LIMIT 10
-        """,
-        (track_id,),
-    )
-    artist_rows = cur.fetchall()
+    artist_rows = get_track_artist_genres(track_id)
     picked = _pick_primary_genre(artist_rows) if artist_rows else None
     if picked:
         picked["source"] = "artist"
@@ -443,11 +354,9 @@ def _resolve_track_genre(cur, track_id: int) -> dict | None:
 @router.get("/api/tracks/{track_id}/genre")
 def api_track_genre_by_id(request: Request, track_id: int):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT 1 FROM library_tracks WHERE id = %s", (track_id,))
-        if not cur.fetchone():
-            return Response(status_code=404)
-        result = _resolve_track_genre(cur, track_id)
+    if not get_track_exists(track_id):
+        return Response(status_code=404)
+    result = _resolve_track_genre(track_id)
     if result is None:
         return {"primary": None, "topLevel": None, "source": None, "preset": None}
     return result
@@ -456,12 +365,10 @@ def api_track_genre_by_id(request: Request, track_id: int):
 @router.get("/api/tracks/by-storage/{storage_id}/genre")
 def api_track_genre_by_storage_id(request: Request, storage_id: str):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT id FROM library_tracks WHERE storage_id = %s", (storage_id,))
-        row = cur.fetchone()
-        if not row:
-            return Response(status_code=404)
-        result = _resolve_track_genre(cur, row["id"])
+    tid = get_track_id_by_storage_id(storage_id)
+    if tid is None:
+        return Response(status_code=404)
+    result = _resolve_track_genre(tid)
     if result is None:
         return {"primary": None, "topLevel": None, "source": None, "preset": None}
     return result
@@ -474,7 +381,6 @@ def api_discover_completeness(request: Request):
     cached = get_cache("discover:completeness", max_age_seconds=86400)
     if cached is not None:
         return cached
-    # No cached data — queue a worker task to compute it
     from crate.db import create_task_dedup
     create_task_dedup("compute_completeness", {})
     return []
@@ -524,23 +430,19 @@ def _stream_file(request: Request, filepath: str):
 @router.get("/api/tracks/{track_id}/stream")
 def api_stream_by_id(request: Request, track_id: int):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT path FROM library_tracks WHERE id = %s", (track_id,))
-        row = cur.fetchone()
-    if not row:
+    path = get_track_path(track_id)
+    if not path:
         return Response(status_code=404)
-    return _stream_file(request, row["path"])
+    return _stream_file(request, path)
 
 
 @router.get("/api/tracks/by-storage/{storage_id}/stream")
 def api_stream_by_storage_id(request: Request, storage_id: str):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT path FROM library_tracks WHERE storage_id = %s", (storage_id,))
-        row = cur.fetchone()
-    if not row:
+    path = get_track_path_by_storage_id(storage_id)
+    if not path:
         return Response(status_code=404)
-    return _stream_file(request, row["path"])
+    return _stream_file(request, path)
 
 
 @router.get("/api/stream/{filepath:path}")
@@ -554,11 +456,9 @@ def api_similar_tracks_query(request: Request, path: str = "", track_id: int = 0
     from crate.bliss import get_similar_from_db
 
     if track_id:
-        with get_db_ctx() as cur:
-            cur.execute("SELECT path FROM library_tracks WHERE id = %s", (track_id,))
-            row = cur.fetchone()
-            if row:
-                path = row["path"]
+        found_path = get_track_path(track_id)
+        if found_path:
+            path = found_path
 
     if not path:
         raise HTTPException(status_code=400, detail="path or track_id required")
@@ -600,23 +500,19 @@ def _download_track(request: Request, filepath: str):
 @router.get("/api/tracks/{track_id}/download")
 def api_download_track_by_id(request: Request, track_id: int):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT path FROM library_tracks WHERE id = %s", (track_id,))
-        row = cur.fetchone()
-    if not row:
+    path = get_track_path(track_id)
+    if not path:
         return Response(status_code=404)
-    return _download_track(request, row["path"])
+    return _download_track(request, path)
 
 
 @router.get("/api/tracks/by-storage/{storage_id}/download")
 def api_download_track_by_storage_id(request: Request, storage_id: str):
     _require_auth(request)
-    with get_db_ctx() as cur:
-        cur.execute("SELECT path FROM library_tracks WHERE storage_id = %s", (storage_id,))
-        row = cur.fetchone()
-    if not row:
+    path = get_track_path_by_storage_id(storage_id)
+    if not path:
         return Response(status_code=404)
-    return _download_track(request, row["path"])
+    return _download_track(request, path)
 
 
 @router.get("/api/download/track/{filepath:path}")
@@ -638,27 +534,26 @@ MOOD_PRESETS = {
 }
 
 
+def _mood_conditions(filters: dict) -> tuple[list[str], list]:
+    conditions = ["bpm IS NOT NULL"]
+    params: list = []
+    for key, val in filters.items():
+        col = key.rsplit("_", 1)[0]
+        op = ">" if key.endswith("_min") else "<"
+        conditions.append(f"{col} {op}= %s")
+        params.append(val)
+    return conditions, params
+
+
 @router.get("/api/browse/moods")
 def api_browse_moods(request: Request):
     """Return available mood presets with track counts."""
     _require_auth(request)
-    from crate.db import get_db_ctx
     results = []
-    with get_db_ctx() as cur:
-        for name, filters in MOOD_PRESETS.items():
-            conditions = ["bpm IS NOT NULL"]
-            params: list = []
-            for key, val in filters.items():
-                col = key.rsplit("_", 1)[0]
-                op = ">" if key.endswith("_min") else "<"
-                conditions.append(f"{col} {op}= %s")
-                params.append(val)
-            cur.execute(
-                f"SELECT COUNT(*) AS cnt FROM library_tracks WHERE {' AND '.join(conditions)}",
-                params,
-            )
-            count = cur.fetchone()["cnt"]
-            results.append({"name": name, "track_count": count, "filters": filters})
+    for name, filters in MOOD_PRESETS.items():
+        conditions, params = _mood_conditions(filters)
+        count = count_mood_tracks(conditions, params)
+        results.append({"name": name, "track_count": count, "filters": filters})
     return results
 
 
@@ -669,27 +564,6 @@ def api_browse_mood_tracks(request: Request, mood: str, limit: int = Query(50, g
     preset = MOOD_PRESETS.get(mood)
     if not preset:
         raise HTTPException(status_code=404, detail=f"Unknown mood: {mood}")
-    from crate.db import get_db_ctx
-    conditions = ["bpm IS NOT NULL"]
-    params: list = []
-    for key, val in preset.items():
-        col = key.rsplit("_", 1)[0]
-        op = ">" if key.endswith("_min") else "<"
-        conditions.append(f"{col} {op}= %s")
-        params.append(val)
-    params.append(limit)
-    with get_db_ctx() as cur:
-        cur.execute(
-            f"""SELECT t.id, t.storage_id, t.title, t.artist, a.name AS album, t.path, t.duration,
-                       ar.id AS artist_id, ar.slug AS artist_slug,
-                       a.id AS album_id, a.slug AS album_slug,
-                       t.bpm, t.energy, t.danceability, t.valence
-                FROM library_tracks t
-                JOIN library_albums a ON a.id = t.album_id
-                LEFT JOIN library_artists ar ON ar.name = t.artist
-                WHERE {' AND '.join(conditions)}
-                ORDER BY RANDOM() LIMIT %s""",
-            params,
-        )
-        tracks = [dict(r) for r in cur.fetchall()]
+    conditions, params = _mood_conditions(preset)
+    tracks = get_mood_tracks(conditions, params, limit)
     return {"mood": mood, "filters": preset, "tracks": tracks, "count": len(tracks)}
