+++ app/tests/test_library_sync.py
@@ -46,16 +46,13 @@ class TestLibrarySyncFullSync:
                  patch("crate.library_sync.upsert_artist") as mock_upsert_artist, \
                  patch("crate.library_sync.upsert_album", return_value=1) as mock_upsert_album, \
                  patch("crate.library_sync.upsert_track") as mock_upsert_track, \
-                 patch("crate.library_sync.get_db_ctx") as mock_ctx, \
+                 patch("crate.library_sync.get_album_id_by_path", return_value=None), \
+                 patch("crate.library_sync.get_tracks_by_album_id", return_value={}), \
+                 patch("crate.library_sync.delete_track_by_path"), \
                  patch("crate.library_sync.delete_artist"), \
                  patch("crate.library_sync.delete_album"), \
                  patch("crate.library_sync.mutagen.File", return_value=None), \
                  patch("crate.library_sync.read_tags", return_value={}):
-                mock_cur = MagicMock()
-                mock_cur.fetchone.return_value = None
-                mock_cur.fetchall.return_value = []
-                mock_ctx.return_value.__enter__ = MagicMock(return_value=mock_cur)
-                mock_ctx.return_value.__exit__ = MagicMock(return_value=False)
 
                 from crate.library_sync import LibrarySync
                 sync = LibrarySync(config)
@@ -99,16 +96,13 @@ class TestLibrarySyncFullSync:
                  patch("crate.library_sync.upsert_artist") as mock_upsert, \
                  patch("crate.library_sync.upsert_album", return_value=1), \
                  patch("crate.library_sync.upsert_track"), \
-                 patch("crate.library_sync.get_db_ctx") as mock_ctx, \
+                 patch("crate.library_sync.get_album_id_by_path", return_value=None), \
+                 patch("crate.library_sync.get_tracks_by_album_id", return_value={}), \
+                 patch("crate.library_sync.delete_track_by_path"), \
                  patch("crate.library_sync.delete_artist"), \
                  patch("crate.library_sync.delete_album"), \
                  patch("crate.library_sync.mutagen.File", return_value=None), \
                  patch("crate.library_sync.read_tags", return_value={}):
-                mock_cur = MagicMock()
-                mock_cur.fetchone.return_value = None
-                mock_cur.fetchall.return_value = []
-                mock_ctx.return_value.__enter__ = MagicMock(return_value=mock_cur)
-                mock_ctx.return_value.__exit__ = MagicMock(return_value=False)
 
                 from crate.library_sync import LibrarySync
                 sync = LibrarySync(config)
@@ -140,14 +134,11 @@ class TestSyncAlbum:
                  patch("crate.library_sync.upsert_artist"), \
                  patch("crate.library_sync.upsert_album", return_value=1), \
                  patch("crate.library_sync.upsert_track") as mock_upsert_track, \
-                 patch("crate.library_sync.get_db_ctx") as mock_ctx, \
+                 patch("crate.library_sync.get_album_id_by_path", return_value=None), \
+                 patch("crate.library_sync.get_tracks_by_album_id", return_value={}), \
+                 patch("crate.library_sync.delete_track_by_path"), \
                  patch("crate.library_sync.mutagen.File", return_value=mock_mf), \
                  patch("crate.library_sync.read_tags", return_value={"artist": "Artist", "album": "Album", "title": "Track"}):
-                mock_cur = MagicMock()
-                mock_cur.fetchone.return_value = None
-                mock_cur.fetchall.return_value = []
-                mock_ctx.return_value.__enter__ = MagicMock(return_value=mock_cur)
-                mock_ctx.return_value.__exit__ = MagicMock(return_value=False)
 
                 from crate.library_sync import LibrarySync
                 sync = LibrarySync(config)
@@ -183,14 +174,11 @@ class TestSyncAlbum:
                  patch("crate.library_sync.upsert_artist"), \
                  patch("crate.library_sync.upsert_album", return_value=1), \
                  patch("crate.library_sync.upsert_track") as mock_upsert_track, \
-                 patch("crate.library_sync.get_db_ctx") as mock_ctx, \
+                 patch("crate.library_sync.get_album_id_by_path", return_value=None), \
+                 patch("crate.library_sync.get_tracks_by_album_id", return_value={}), \
+                 patch("crate.library_sync.delete_track_by_path"), \
                  patch("crate.library_sync.mutagen.File", return_value=mock_mf), \
                  patch("crate.library_sync.read_tags", return_value={"artist": "Artist", "album": "Album Deluxe", "title": "Track"}):
-                mock_cur = MagicMock()
-                mock_cur.fetchone.return_value = None
-                mock_cur.fetchall.return_value = []
-                mock_ctx.return_value.__enter__ = MagicMock(return_value=mock_cur)
-                mock_ctx.return_value.__exit__ = MagicMock(return_value=False)
 
                 from crate.library_sync import LibrarySync
                 sync = LibrarySync(config)
@@ -251,24 +239,35 @@ class TestRemoveStale:
                 "audio_extensions": [".flac"],
             }
 
-            with patch("crate.library_sync.get_db_ctx") as mock_ctx, \
+            artists_data = [
+                {"name": "Existing Artist", "folder_name": "Existing Artist", "album_count": 1, "track_count": 5,
+                 "total_size": 0, "formats": [], "primary_format": None, "has_photo": 0, "dir_mtime": None,
+                 "updated_at": None, "id": 1, "storage_id": "fake-uuid-1", "slug": "existing-artist",
+                 "bio": None, "tags_json": None, "similar_json": None, "spotify_id": None,
+                 "spotify_popularity": None, "mbid": None, "country": None, "area": None,
+                 "formed": None, "ended": None, "artist_type": None, "members_json": None,
+                 "urls_json": None, "listeners": None, "enriched_at": None, "discogs_id": None,
+                 "spotify_followers": None, "lastfm_playcount": None, "discogs_profile": None,
+                 "discogs_members_json": None, "latest_release_date": None, "content_hash": None,
+                 "formats_json": "[]"},
+                {"name": "Gone Artist", "folder_name": "Gone Artist", "album_count": 1, "track_count": 3,
+                 "total_size": 0, "formats": [], "primary_format": None, "has_photo": 0, "dir_mtime": None,
+                 "updated_at": None, "id": 2, "storage_id": "fake-uuid-2", "slug": "gone-artist",
+                 "bio": None, "tags_json": None, "similar_json": None, "spotify_id": None,
+                 "spotify_popularity": None, "mbid": None, "country": None, "area": None,
+                 "formed": None, "ended": None, "artist_type": None, "members_json": None,
+                 "urls_json": None, "listeners": None, "enriched_at": None, "discogs_id": None,
+                 "spotify_followers": None, "lastfm_playcount": None, "discogs_profile": None,
+                 "discogs_members_json": None, "latest_release_date": None, "content_hash": None,
+                 "formats_json": "[]"},
+            ]
+
+            with patch("crate.library_sync.get_library_artists", return_value=(artists_data, 2)), \
                  patch("crate.library_sync.get_library_artist", return_value=None), \
+                 patch("crate.library_sync.get_album_paths_for_artist", return_value=[]), \
+                 patch("crate.library_sync.get_all_album_paths", return_value=[]), \
                  patch("crate.library_sync.delete_artist") as mock_delete, \
                  patch("crate.library_sync.delete_album"):
-                mock_cur = MagicMock()
-                # First call: artists query returns existing + stale (with folder_name, album_count, track_count)
-                # Second call: albums for stale artist (empty paths)
-                # Third call: albums query for remove_stale albums phase
-                mock_cur.fetchall.side_effect = [
-                    [
-                        {"name": "Existing Artist", "folder_name": "Existing Artist", "album_count": 1, "track_count": 5},
-                        {"name": "Gone Artist", "folder_name": "Gone Artist", "album_count": 1, "track_count": 3},
-                    ],
-                    [],  # album paths for "Gone Artist"
-                    [],  # albums phase
-                ]
-                mock_ctx.return_value.__enter__ = MagicMock(return_value=mock_cur)
-                mock_ctx.return_value.__exit__ = MagicMock(return_value=False)
 
                 from crate.library_sync import LibrarySync
                 sync = LibrarySync(config)
